# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API SIHIOS-TS",
        "description": "API de la plateforme SIHIOS-TS",
        "contact": {
            "name": "Developer",
            "email": "michel.banh@softskills.ci"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/forgot-password": {
            "post": {
                "tags": [
                    "Réinitialisation de Mot de Passe"
                ],
                "summary": "Envoyer un lien de réinitialisation de mot de passe",
                "operationId": "64f53b8c0d28e63057ec35a76901ce44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "utilisateur@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien de réinitialisation envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de l'envoi du lien",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Réinitialisation de Mot de Passe"
                ],
                "summary": "Réinitialiser le mot de passe",
                "operationId": "f66d3cf90ea9cc52c86d7f415f2c1525",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "token_de_réinitialisation"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "utilisateur@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nouveauMotDePasse123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nouveauMotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de la réinitialisation du mot de passe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}